Moved code to defcon source directory.
