+++ defcon/_augmentors.py
