--- leb/__init__.py
@@ -1,3 +1,0 @@
-# © All rights reserved. ECOLE POLYTECHNIQUE FEDERALE DE LAUSANNE,
-# Switzerland, Laboratory of Experimental Biophysics
-# See the LICENSE.txt file for more details.
